filepathfilter: always use Git-compatible pattern matching
Currently, we have some custom code that's designed to make pattern
matching work in a way that's compatible with an older, broken
implementation of wildmatch.  Unfortunately, this leads to a variety of
unexpected behavior that's hard to reason about.  Let's instead switch
to be compatible with the behavior of Git's .gitattributes file in all
cases, so that the behavior is standard and documented.

In order to do so, set the Basename flag, which provides the behavior we
want from .gitattributes, where patterns without a slash end up matching
a file in any directory.  Remove the special-casing we have for
non-strict mode and always use strict mode, and drop the now-useless
dirs flag.  Update the tests to reflect all of the patterns that have
changed.